+++ packages/js/src/Modules/Verto/index.ts
@@ -34,7 +34,10 @@ export default class Verto extends BrowserSession {
         Object.keys(this.calls).forEach((callId) => {
           if (this.calls[callId]) {
             logger.info(`Hanging up call due to window unload: ${callId}`);
-            void this.calls[callId].hangup({}, true);
+            void this.calls[callId].hangup(
+              { initiator: 'sdk:beforeunload' },
+              true
+            );
           }
         });
       }